Seperated Toys and VibeLobbies, also made UI progress for the two windows. - Migrated Intiface Central from the old controller to an IpcCaller - Restructured BuzzToyDevice control to follow the same structure as the State folder. - Added BuzzToyFileSystem & Selector

+++ ProjectGagSpeak/Services/Mediator/Messages/Messages.UI.cs
@@ -82,7 +82,7 @@ public record VerificationPopupMessage(VerificationCode VerificationCode) : Mess
 /// <summary> Fires whenever we finished recording a new pattern, and need to finalize it's details. </summary>
 /// <param name="StoredData"> The data we are saving. </param>
 /// <param name="Duration"> The duration of the pattern. </param>
-public record PatternSavePromptMessage(List<byte> StoredData, TimeSpan Duration) : MessageBase;
+public record PatternSavePromptMessage(List<double> StoredData, TimeSpan Duration) : MessageBase;
 
 
 public record ClosePatternSavePromptMessage : MessageBase;
